feat: add support for useSafeArea and for embedding outside of a dialog
In order to fill the whole screen with an image, the `useSafeArea` option
of the `showDialog` call can now be passed through, and it defaults to false.

Also, the size of `EasyImageView` is now not hard-coded to the size of the screen,
but instead uses `SizedBox.expand` to fill its parent. This enables embedding
of `EasyImageView` and `EasyImageViewPager` inside of another widget, without
using a `Dialog`.

Also adds an example of how to do that.