Enable dragging the PageView with a mouse (GH-11)
By default page views (and scroll views in general) can only be
dragged with touch devices, not mice.
That means it was impossible to advance to the next image on desktop
(Windows).
This commit enables dragging with a mouse. Since the images are
displayed full-screen, it shouldn't interfere with anything else.

For more references, see:
https://stackoverflow.com/questions/69424933/flutter-pageview-not-swipeable-on-web-desktop-mode
https://docs.flutter.dev/release/breaking-changes/default-scroll-behavior-drag
https://www.reddit.com/r/flutterhelp/comments/thyl47/how_to_use_arrow_keys_or_mouse_swipe_for_pageview/