feat: Add option to allow swipe-down to dismiss
An ticket asked for swipe-to-dismiss (https://github.com/thesmythgroup/easy_image_viewer/issues/14)
It's easy to implement and now supported with the `swipeDismissable` argument.
This allows the user to drag the dialog down to dismiss it.
However, this is still WIP because when the user has zoomed in, you can't pan the
image anymore because the Dismissable widget begins to dismiss. Will have to address
this in a future commit.