Demonstration of an issues with PageController "losing" its page
I'm trying to implement a solution for swipe-to-dismiss using Dismissable.
It's kind of working, but I had to activate/deactivate the Dismissable based on
whether the user has zoomed in on an image or not. Here's the problem: When I zoom in,
the PageView jumps back to the first image.