feat: double-tap to zoom - make double tap work with manual zooming
Double tap didn't work correctly when combined with manual zooming.
This commit fixes that by taking into account the current zoom scale
when deciding how to handle a double tap. It also ensures that the
`onScaleChanged` callback is called.

Tests still fail and additional tests still need to be written for the
new behavior.