GH #34: Initial work to add loading indicator
This adds initial support for customizing the Image widget that is being displayed.
That includes customizing Image's `frameBuilder` and `loadingBuilder`. Rather
than providing those callbacks as parameters, it's more versatile to let the
user create their own Image widget using their desired configuration.

This is now supported in `EasyImageProvider` through a new `imageWidgetBuilder` method.
The default implementation shows a loading indicator and does a quick fade-in animation
once the image is loaded.

By using your own EasyImageProvider subclass this can be fully customized. Some tests
still fail and more tests need to be added, but this is a first shot at this.