Fix: Revert breaking change from WillPopScope to PopScope to avoid higher Flutter version
In 1.3.0 a deprecation warning was addressed by changing WillPopScope to PopScope in one
place. However, PopScope requires Flutter 3.13.0 and I had omitted to bump the minimum
Flutter version, which would be a breaking change. To avoid that, roll back the change
and ignore the deprecation warning until we release a major version.