Default barrierColor to be same as backgroundColor when useSafeArea: true
Improve SafeArea support by defaulting the `barrierColor` to be the same as the `backgroundColor`, thus eliminating visually ugly bars on the top and bottom when using `useSafeArea: true`. The `barrierColor` can be now also be set explicitly (see [GH-62](https://github.com/thesmythgroup/easy_image_viewer/issues/62))

Also improve the exmaple app: It's not embedded in a SafeArea and SingleChildScrollView to fix overflow layout errors on smaller devices.