Default barrierColor to be same as backgroundColor when useSafeArea: true (#63)
* Default barrierColor to be same as backgroundColor when useSafeArea: true

Improve SafeArea support by defaulting the `barrierColor` to be the same as the `backgroundColor`, thus eliminating visually ugly bars on the top and bottom when using `useSafeArea: true`. The `barrierColor` can be now also be set explicitly (see [GH-62](https://github.com/thesmythgroup/easy_image_viewer/issues/62))

Also improve the exmaple app: It's not embedded in a SafeArea and SingleChildScrollView to fix overflow layout errors on smaller devices.

* ci: Update setup-java and checkout actions to v4

Pipeline was failing because Node 16 actions are deprecated. So the two
actions had to be updated to v4.

* ci: Specify distribution for new actions/setup-java version

Starting with v2, you need to specify the distribution. v1 used "zulu", so
we stick with the same.